fix: 🐛 update phone number to type string
